refactor(server): make `closeSession` db operations reusable
