+++ packages/server/src/db/mutation/session-mutation.ts
@@ -2,11 +2,10 @@
 import type { GameSession, SessionStatus } from "@repo/db"
 import type { ClientSessionVariants } from "@repo/schema/session"
 
-// helpers
-import { calculateElo } from "@repo/helper/elo"
-
-// server
+// db
 import { db } from "@/db"
+import { playerStatsUpdaterOperations } from "@/db/transaction/player-transaction"
+import { closeSessionOperation } from "@/db/transaction/session-transaction"
 
 /**
  * Closes a game session and updates player statistics.
@@ -16,61 +15,20 @@ import { db } from "@/db"
  * - Records the session results for each participating player.
  * - Operations are executed within a database transaction.
  * 
- * @param {ClientSession} clientSession The current session containing player data, mode, and stats.
+ * @param {ClientSessionVariants} session The current session containing player data, mode, and stats.
  * @param {string} requesterPlayerId The ID of the player requesting the session closure.
  * @param {SessionStatus} status Session "action" status.
  * @returns {Promise<GameSession>} The updated game session.
  */
 export async function closeSession(
-  clientSession: Pick<ClientSessionVariants, "slug" | "mode" | "format" | "tableSize" | "owner" | "guest" | "cards" | "stats">,
+  session: Pick<ClientSessionVariants, "slug" | "mode" | "format" | "tableSize" | "owner" | "guest" | "cards" | "stats">,
   requesterPlayerId: string,
   status: Extract<SessionStatus, "FINISHED" | "CLOSED" | "FORCE_CLOSED">
 ): Promise<GameSession> {
-  const { slug, format, owner, guest, cards, stats } = clientSession
-
-  const players = [owner]
-  if (format === "PVP" || format === "COOP") players.push(guest!)
-
-  /* Updates the statistics of session players */
-  const playerStatsOperations = players.map(
-    (player) => db.playerProfile.update({
-      where: { id: player.id },
-      data: {
-        stats: {
-          elo: calculateElo(clientSession, player.id, status, requesterPlayerId).newElo,
-          flips: player.stats.flips + stats.flips[player.id],
-          matches: player.stats.matches + stats.matches[player.id],
-          timer: player.stats.timer + stats.timer,
-          sessions: ++player.stats.sessions
-        }
-      }
-    })
-  )
-
-  const [session] = await Promise.all([
-    /* Updates the affected game session based on the given status */
-    db.gameSession.update({
-      where: { slug },
-      data: {
-        cards,
-        stats,
-        status,
-        closedAt: new Date(),
-        results: {
-          createMany: {
-            data: players.map((player) => ({
-              playerId: player.id,
-              gainedElo: calculateElo(clientSession, player.id, status, requesterPlayerId).gainedElo,
-              flips: stats.flips[player.id],
-              matches: stats.matches[player.id],
-              timer: stats.timer
-            }))
-          }
-        }
-      }
-    }),
-    db.$transaction(playerStatsOperations)
+  const [closedSession] = await Promise.all([
+    closeSessionOperation(session, requesterPlayerId, status),
+    db.$transaction(playerStatsUpdaterOperations(session, requesterPlayerId, status))
   ])
 
-  return session
+  return closedSession
 }
